Added version moved test
