+++ compatibility.json
@@ -1,5 +1,7 @@
 {
   "rpcChainVMProtocolVersion": {
+    "v0.4.2": 18,
+    "v0.4.1": 18,
     "v0.4.0": 17,
     "v0.3.0": 16,
     "v0.2.9": 15,
