+++ plugin/evm/version_test.go
@@ -1,7 +1,8 @@
-package utils
+package evm
 
 import (
 	"encoding/json"
+	"fmt"
 	"os"
 	"testing"
 
@@ -16,11 +17,25 @@ func TestCompatibility(t *testing.T) {
 	subevmVersion := "v0.4.0"
 	expectedRPCVersion := 17
 
-	compat, err := os.ReadFile("../compatibility.json")
+	compat, err := os.ReadFile("../../compatibility.json")
 	assert.NoError(t, err)
 
 	var parsedCompat rpcChainCompatibility
 	err = json.Unmarshal(compat, &parsedCompat)
 	assert.NoError(t, err)
 	assert.Equal(t, expectedRPCVersion, parsedCompat.RPCChainVMProtocolVersion[subevmVersion])
+}
+
+func TestCompatibilityCurrentVersion(t *testing.T) {
+	compat, err := os.ReadFile("../../compatibility.json")
+	assert.NoError(t, err)
+
+	var parsedCompat rpcChainCompatibility
+	err = json.Unmarshal(compat, &parsedCompat)
+	assert.NoError(t, err)
+
+	fmt.Println("Version:", Version)
+
+	_, valueInJSON := parsedCompat.RPCChainVMProtocolVersion[Version]
+	assert.True(t, valueInJSON)
 }
